+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Unreleased
 ## Added
+- `time_residuals`, `scaled_toa_uncertainties`, and `model_dm`, `from_pint` methods in `SPNTA`
 ## Changed
 ## Fixed
 ## Removed
+- `test_alloc` and `test_likelihood` from `test_data_files.py`
 
 # 0.0.6
 ## Added